backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
alembic==1.13.3
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic[email]==2.9.2